Adding Required Files/Headers for GSHM GitHub Repo
Summary:
We will use ShipIt to sync our GitHub repo for the GSHM paper with files
that will be stored in fbcode. This requires syncing all files that will be on the
GitHub repo with files that are in fbcode. The preliminary ShipIt config has synced
the paper R code file already:
https://github.com/facebookresearch/gaussian_sparse_histogram_mechanism

Reviewed By: bkarrer

Differential Revision: D35104949

fbshipit-source-id: 42334977262df0dd959ee635700aaa809f92d8a9